refactor: reorganize imports and improve code quality
