+++ src/domain/contracts/repositories/pokemon/index.ts
@@ -1,3 +1,3 @@
-export * from './list-pokemon.repository'
 export * from './find-pokemon-by-id.repository'
 export * from './find-pokemon-by-type.repository'
+export * from './list-pokemon.repository'
