+++ src/domain/services/pokemon/index.ts
@@ -1,3 +1,3 @@
-export * from './list-pokemon.service'
 export * from './find-pokemon-by-id.service'
 export * from './find-pokemon-by-type.service'
+export * from './list-pokemon.service'
