+++ src/main/factories/controllers/index.ts
@@ -1,2 +1,2 @@
-export * from './list-pokemon.controller.factory'
 export * from './find-pokemon-by-id.controller.factory'
+export * from './list-pokemon.controller.factory'
